SIIM/requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.0
altair==4.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
audiomentations==0.27.0
audioread==3.0.0
Automat==0.8.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.4b2
bleach==5.0.1
blessed==1.19.1
blinker==1.4
blis==0.7.8
bounding-box==0.1.3
cachetools==5.2.0
catalogue==2.0.8
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
cloud-init==22.2
cloudpickle==2.1.0
cmake==3.24.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
ctcdecode==1.0.3
cycler==0.11.0
cymem==2.0.6
cytoolz==0.12.0
d2l==0.17.5
datasets==2.1.0
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.6+cu102
dill==0.3.5.1
distro==1.4.0
distro-info===0.23ubuntu1
docker==5.0.3
docker-pycreds==0.4.0
easybuild-easyblocks==4.5.4
easybuild-easyconfigs==4.5.4
easybuild-framework==4.5.4
easydict==1.10
effdet==0.2.3
efficientnet-pytorch==0.7.1
entrypoints==0.3
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
filterpy==1.4.5
flatbuffers==2.0.7
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.0.0
GPUtil==1.4.0
grpcio==1.47.0
h5py==3.7.0
httplib2==0.14.0
huggingface-hub==0.5.1
hydra-core==1.2.0
hyperlink==19.0.0
idna==2.8
imageio==2.21.1
imagesize==1.4.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
imutils==0.5.4
incremental==16.10.1
iopath==0.1.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.3
langcodes==3.3.0
language-selector==0.1
lap==0.4.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.20.5
libclang==14.0.6
librosa==0.9.1
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.4
more-itertools==4.2.0
motmetrics==1.2.5
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
munch==2.5.0
murmurhash==1.0.8
mypy-extensions==0.4.3
mysqlclient==1.4.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8.5
ninja==1.10.2.3
nltk==3.7
notebook==6.4.12
numba==0.56.2
numpy==1.22.0
nvidia-ml-py==11.495.46
oauthlib==3.1.0
omegaconf==2.2.3
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.1
onnxruntime==1.8.0
opencv-python==4.6.0.66
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
packaging==21.3
pandarallel==1.6.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pathy==0.6.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.6.0
portalocker==2.5.1
preshed==3.0.7
pretrainedmodels==0.7.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-vncorenlp==0.1.3
pyarrow==9.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocotools==2.0.4
pycodestyle==2.9.1
pycosat==0.6.3
pycparser==2.21
pydantic==1.9.2
pyDeprecate==0.3.2
pydicom==2.1.1
pydot==1.4.2
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
pyjnius==1.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-Levenshtein==0.20.5
python-magic==0.4.27
python-slugify==7.0.0
pytorch-lightning==1.7.2
pytorch-metric-learning==1.6.2
pytorch-warmup==0.1.0
pytz==2022.1
pyvi==0.1.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.2.0
qudida==0.0.4
ranger-fm==1.9.3
rapidfuzz==2.11.1
regex==2022.7.25
requests==2.25.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resampy==0.4.2
responses==0.18.0
retrying==1.3.4
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3cmd==2.3.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
segmentation-models-pytorch==0.3.1
Send2Trash==1.8.0
sentry-sdk==1.9.5
service-identity==18.1.0
setproctitle==1.3.2
shortuuid==1.0.9
simplejson==3.16.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
smmap==5.0.0
sos==4.3
soundfile==0.11.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
ssh-import-id==5.10
stack-data==0.3.0
sympy==1.10.1
systemd-python==234
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.1
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
thinc==8.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchdata==0.4.1
torchmetrics==0.9.3
torchtext==0.13.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
Twisted==18.9.0
typer==0.4.2
typing-extensions==4.3.0
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
wadllib==1.3.3
wandb==0.13.2
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.1.2
widgetsnbextension==4.0.2
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.8.1
youtokentome==1.0.6
zipp==3.8.1
zope.interface==4.7.1
